+++ lib/glimmer/command_handler_chain_link.rb
@@ -16,7 +16,8 @@ module Glimmer
         # TODO see if we need a better response here (e.g. dev mode error raising vs production mode silent failure)
         # Glimmer.logger.error "Command: #{command_symbol} cannot be handled!"
         Glimmer.logger.debug "Command: #{command_symbol} cannot be handled!"
-        raise "Command: #{command_symbol} cannot be handled!"
+        # raise "Command: #{command_symbol} cannot be handled!"
+        nil
       end
     end
   end